+++ pkg/logger/field_test.go
@@ -110,7 +110,7 @@ func TestFieldAddingError(t *testing.T) {
 		{t: ArrayMarshalerType, iface: users(-1), want: []interface{}{}, err: "too few users"},
 		{t: ObjectMarshalerType, iface: users(-1), want: map[string]interface{}{}, err: "too few users"},
 		{t: InlineMarshalerType, iface: users(-1), want: nil, err: "too few users"},
-		{t: StringerType, iface: obj{}, want: empty, err: "PANIC=interface conversion: logger.obj is not fmt.Stringer: missing method String"},
+		{t: StringerType, iface: obj{}, want: empty, err: "PANIC=interface conversion: global.obj is not fmt.Stringer: missing method String"},
 		{t: StringerType, iface: &obj{1}, want: empty, err: "PANIC=panic with string"},
 		{t: StringerType, iface: &obj{2}, want: empty, err: "PANIC=panic with reason"},
 		{t: StringerType, iface: &obj{3}, want: empty, err: "PANIC=<nil>"},
@@ -121,7 +121,7 @@ func TestFieldAddingError(t *testing.T) {
 		enc := NewMapObjectEncoder()
 		assert.NotPanics(t, func() { f.AddTo(enc) }, "Unexpected panic when adding fields returns an reason.")
 		assert.Equal(t, tt.want, enc.Fields["k"], "On reason, expected zero value in field.Key.")
-		assert.Equal(t, tt.err, enc.Fields["kError"], "Expected reason message in logger context.")
+		assert.Equal(t, tt.err, enc.Fields["kError"], "Expected reason message in global context.")
 	}
 }