+++ pkg/logger/global_test.go
@@ -32,10 +32,10 @@ func Test_logger(t *testing.T) {
 
 	Error("Error", "Error")
 	Errorf("this %s", "Errorf")
-	helper.GetRaw().(*StdLogger).OpenMsgFormat()
-	DebugFiled("access", String("fd", "fd"), Int("fd", 456))
-	InfoFiled("access", String("fd", "fd"), Int("fd", 456))
-	ErrorFiled("access", String("fd", "fd"), Int("fd", 456), Err(errors.New("fdasfdf")))
+	DebugField("access", String("fd", "fd"), Int("fd", 456))
+	InfoField("access", String("fd", "fd"), Int("fd", 456))
+	ErrorField("access", String("fd", "fd"), Int("fd", 456), Err(errors.New("fdasfdf")))
+
 }
 
 func Test(t *testing.T) {
@@ -44,4 +44,25 @@ func Test(t *testing.T) {
 		Debugf("free: %.3f", time.Since(start).Seconds())
 	}()
 	time.Sleep(time.Millisecond * 3)
+}
+
+func Test_WithFields(t *testing.T) {
+	lg := WithFields(String("mod", "test"))
+	lg.Debug("Debug")
+	lg.Info("Info")
+	lg.Warn("Warn")
+	lg.Error("Error")
+	//lg.Fatal("Fatal")
+
+	lg.Debugf("Debug %s", "test")
+	lg.Infof("Info %s", "test")
+	lg.Warnf("Warn %s", "test")
+	lg.Errorf("Error %s", "test")
+	//lg.Fatalf("Fatal %s", "test")
+
+	lg.DebugField("Debug", String("k", "test"))
+	lg.InfoField("Info", String("k", "test"))
+	lg.WarnField("Warn", String("k", "test"))
+	lg.ErrorField("Error", String("k", "test"))
+	lg.FatalField("Fatal", String("k", "test"))
 }