+++ pkg/logger/level.go
@@ -1,17 +1,3 @@
-// Copyright 2022 The imkuqin-zw Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 package logger
 
 import (
@@ -21,7 +7,7 @@ import (
 
 type Level int8
 
-// String returns a lower-case ASCII representation of the logger level.
+// String returns a lower-case ASCII representation of the global level.
 func (l Level) String() string {
 	switch l {
 	case LvDebug:
@@ -31,7 +17,7 @@ func (l Level) String() string {
 	case LvWarn:
 		return "warn"
 	case LvError:
-		return "reason"
+		return "error"
 	case LvFault:
 		return "fatal"
 	default:
@@ -79,27 +65,10 @@ func (l *Level) unmarshalText(text []byte) bool {
 	return true
 }
 
-type Logger interface {
-	Debug(args ...interface{})
-	Info(args ...interface{})
-	Warn(args ...interface{})
-	Error(args ...interface{})
-	Fatal(args ...interface{})
-
-	Debugf(format string, args ...interface{})
-	Infof(format string, args ...interface{})
-	Warnf(format string, args ...interface{})
-	Errorf(format string, args ...interface{})
-	Fatalf(format string, args ...interface{})
-
-	Debugw(msg string, kvs ...interface{})
-	Infow(msg string, kvs ...interface{})
-	Warnw(msg string, kvs ...interface{})
-	Errorw(msg string, kvs ...interface{})
-	Fatalw(msg string, kvs ...interface{})
-
-	SetLevel(Level)
-	GetLevel() Level
-	Enable(Level) bool
-	Clone() Logger
-}
+const (
+	LvDebug Level = iota - 1
+	LvInfo
+	LvWarn
+	LvError
+	LvFault
+)